+++ src/features/reader/viewer/ReaderViewer.tsx
@@ -230,6 +230,7 @@ const BaseReaderViewer = ({
         pageScaleMode,
         shouldStretchPage,
         readerWidth,
+        themeDirection,
     );
 
     useLayoutEffect(() => {